Rename `FieldTypes` to `FieldTypesProto`
